+++ Sources/KSPlayer/MEPlayer/KSMEPlayer.swift
@@ -356,7 +356,7 @@ extension KSMEPlayer: MediaPlayerProtocol {
 
     public var duration: TimeInterval { playerItem.duration }
 
-    public var fileSize: Double { playerItem.fileSize }
+    public var fileSize: Int64 { playerItem.fileSize }
 
     public var seekable: Bool { playerItem.seekable }
